+++ core-ng/src/main/java/core/framework/impl/web/request/RequestImpl.java
@@ -2,6 +2,7 @@ package core.framework.impl.web.request;
 
 import core.framework.http.ContentType;
 import core.framework.http.HTTPMethod;
+import core.framework.impl.validate.ValidationException;
 import core.framework.impl.web.bean.RequestBeanMapper;
 import core.framework.util.Encodings;
 import core.framework.util.Exceptions;
@@ -14,7 +15,6 @@ import core.framework.web.exception.BadRequestException;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.server.handlers.Cookie;
 
-import java.io.UncheckedIOException;
 import java.util.Map;
 import java.util.Optional;
 
@@ -155,8 +155,8 @@ public final class RequestImpl implements Request {
             } else {
                 throw Exceptions.error("not supported method, method={}", method);
             }
-        } catch (UncheckedIOException e) {
-            throw new BadRequestException(e.getMessage(), BadRequestException.DEFAULT_ERROR_CODE, e);
+        } catch (ValidationException e) {
+            throw new BadRequestException(e.getMessage(), "VALIDATION_ERROR", e);
         }
     }
 }